% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crda.classify.R
\name{crda.classify}
\alias{crda.classify}
\title{Classification example by CRDA}
\usage{
crda.classify(L = 10, nK = 10, q = Inf, prior = "uniform",
  centerX = FALSE, comparison = FALSE)
}
\arguments{
\item{L}{Number of training and test splits}

\item{nK}{Number of candidates in CV grid for joint-sparsity level}

\item{q}{Type of Lq,1 norm}

\item{prior}{Type of prior class probabilities}

\item{centerX}{Flag for centering the training dataset}

\item{comparison}{Flag for comparison with other methods}
}
\description{
The function \code{crda.cv} performs Classification by CRDA-variants
}
\details{
crda.classify
}
\examples{
crda.classify()
}
\seealso{
\code{\link{crda}}, \code{\link{crda.setup3}}, \code{\link{crda.regparam}}, \code{\link{crda.cv}}
}
\author{
Muhammad Naveed Tabassum and Esa Ollila, 2018
}
