% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crda.cv.R
\name{crda.cv}
\alias{crda.cv}
\title{Cross-validation for CRDA}
\usage{
crda.cv(X, y, q = Inf, al = NULL, prior = "uniform", flds = NULL,
  nFolds = 5, Kgrid = round(exp(seq(log(0.05 * nrow(X)), log(0.5 *
  nrow(X)), length.out = 10))), centerX = FALSE, plotCV = FALSE)
}
\arguments{
\item{X}{Training dataset}

\item{y}{Labels for training dataset}

\item{q}{Type of Lq,1 norm}

\item{al}{regularization parameter}

\item{prior}{Type of prior class probabilities}

\item{flds}{Folds for cross-validation}

\item{nFolds}{number of folds}

\item{Kgrid}{A grid having candidate values of joint-sparsity parameter}

\item{centerX}{Flag for centering the training dataset}

\item{plotCV}{Flag for plotting the results}
}
\value{
An object of class \code{crda.cv} with the following attributes
\item{funCall}{The call for CRDA}
}
\description{
The function \code{crda.cv} performs cross-validation
}
\details{
crda.cv
}
\examples{
crda.cv(X,y) when using defaults
crda.cv(X,y, q=1,nFolds=10,centerX=TRUE)
}
\seealso{
\code{\link{crda.regparam}}, \code{\link{crda}}
}
\author{
Muhammad Naveed Tabassum and Esa Ollila, 2018
}
