% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crda.demo.R
\name{crda.demo}
\alias{crda.demo}
\title{Classification Example (demo) for Compressive Regularized Discriminant Analysis (CRDA) approach using the Ell2-RSCM estimator of the covariance matrix.}
\usage{
crda.demo(prior = "uniform")
}
\value{
An object \code{res} of class \code{crda.demo} with the following attributes:
\item{funCall}{The call to the \code{crda.demo} function.}
\item{TER}{Test error rate (TER)}
\item{FSR}{Feature selection rate (FSR)}
\item{CT}{Average computational time (CT) in seconds.}
}
\description{
The function \code{crda.demo} performs classification using CRDA2 for one split of a real genomic dataset, Khan'2001.
}
\details{
crda.demo
}
\examples{
crda.demo()
crda.demo(prior = 'estimated')
}
\seealso{
\code{\link{crda}}, \code{\link{crda.auxfuns}}
}
\author{
Muhammad Naveed Tabassum and Esa Ollila, 2019.
}
