% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crda.setup3.R
\name{crda.setup3}
\alias{crda.setup3}
\title{Generation of Partially Synthetic Dataset}
\usage{
crda.setup3()
}
\value{
An object \code{dataset} of class \code{crda.setup3} with the following attributes:
\item{X}{Training dataset, a pxn matrix with n-samples each having p-features.}
\item{y}{Labels for training dataset, an nx1 vector of whole numbers.}
\item{Xt}{Test dataset, a pxnt matrix with nt-samples each having p-features.}
\item{yt}{Labels for test dataset, an ntx1 vector of whole numbers.}
\item{truePos}{True positions of relevant, i.e., differentially expressed features.}
}
\description{
This function generates and splits partially synthetic data into training and test sets for setup 3 of the CRDA-paper.
}
\details{
crda.setup3
}
\examples{
crda.setup3()
}
