% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crda.auxfuns.R
\name{CRDA0}
\alias{CRDA0}
\alias{crda.auxfuns}
\title{Auxiliary functions used by the CRDA approach}
\usage{
CRDA0(Xt = NULL, X, y, K = NULL, q = "inf",
  prior = table(y)/base::length(y), B = NULL, M = NULL, Ind = NULL)
}
\description{
crda.auxfuns
}
\seealso{
\code{\link{crda}}, \code{\link{crda.demo}}
}
\author{
Muhammad Naveed Tabassum and Esa Ollila, 2019.
}
