% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crda.cv.R
\name{crda.cv}
\alias{crda.cv}
\title{Cross-validation based Joint-sparsity Level for CRDA Method}
\usage{
crda.cv(X, y, q = Inf, al = NULL, prior = "uniform", flds = NULL,
  nFolds = 5, Kgrid = round(exp(seq(log(0.05 * nrow(X)), log(0.5 *
  nrow(X)), length.out = 10))), centerX = FALSE, plotCV = FALSE)
}
\arguments{
\item{X}{Training dataset, a pxn matrix with n-samples each having p-features.}

\item{y}{Labels for training dataset, an nx1 vector of whole numbers.}

\item{q}{Type of Lq,1 norm, default is Linf-norm.}

\item{al}{Regularization parameter.}

\item{prior}{Type of prior class probabilities, either 'uniform' (default) or 'estimated'.}

\item{flds}{Folds for cross-validation (CV).}

\item{nFolds}{number of folds.}

\item{Kgrid}{A grid having candidate values of joint-sparsity level.}

\item{centerX}{Flag for grand-mean centering of test dataset using grand-mean of training dataset.}

\item{plotCV}{Flag for plotting the CV-results.}
}
\value{
An object \code{cv.obj} of class \code{crda.cv} with the following attributes:
\item{funCall}{The call to the \code{crda.cv} function.}
\item{Kgrid}{A grid having candidate values of joint-sparsity level.}
\item{folds}{Folds for cross-validation (CV).}
\item{cvMER}{CV misclassification error.}
\item{resubMER}{Resubstitution error.}
\item{K}{CV-estimate of joint-sparsity level.}
\item{regparam}{The value of regularization parameter.}
}
\description{
The function \code{crda.cv} performs cross-validation for finding the joint-sparsity level of CRDA method.
}
\details{
crda.cv
}
\examples{
crda.cv(X,y)
crda.cv(X,y, q = 1, nFolds = 10)
crda.cv(X,y, q = 1, prior = 'estimated')
crda.cv(X,y, q = 1, centerX = TRUE)
}
\seealso{
\code{\link{crda.regparam}}, \code{\link{crda}}
}
\author{
Muhammad Naveed Tabassum and Esa Ollila, 2018
}
