% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crda.demo.R
\name{crda.demo}
\alias{crda.demo}
\title{A demo (example) on classification of a real genomic dataset}
\usage{
crda.demo(prior = "uniform")
}
\value{
An object \code{res} of class \code{crda.demo} with the following attributes:
\item{funCall}{The call to the \code{crda.demo} function.}
\item{TER}{Test error rate (TER) in terms of percentage of misclassifications.}
\item{FSR}{Feature selection rate (FSR) in terms of percentage of features used.}
\item{CT}{Computational time (CT) in seconds.}
}
\description{
The function \code{crda.demo} performs classification using compressive regularized
discriminant analysis (CRDA) approach for one split of a real genomic dataset, Khan'2001.
}
\details{
crda.demo
}
\examples{
crda.demo()
crda.demo(prior = 'estimated')
}
\seealso{
\code{\link{crda}}, \code{\link{crda.auxfuns}}
}
\author{
Muhammad Naveed Tabassum and Esa Ollila, 2019.
}
