% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crda.R
\name{crda}
\alias{crda}
\title{Classification approach}
\usage{
crda(X, y, q = Inf, al = NULL, K = NULL, Xt = NULL,
  prior = "uniform", centerX = FALSE)
}
\arguments{
\item{X}{Training dataset}

\item{y}{Labels for training dataset}

\item{q}{Type of Lq,1 norm, default is Linf-norm.}

\item{al}{regularization parameter}

\item{K}{Joint-sparsity parameter}

\item{Xt}{Test dataset}

\item{prior}{Type of prior class probabilities}

\item{centerX}{Flag for grand-mean centering of test dataset using grand-mean of training dataset}
}
\value{
An object of class \code{crda} with the following attributes
\item{funCall}{The call for CRDA}
\item{prior}{Prior class probabilities on the training dataset \code{X}.}
}
\description{
The function \code{crda} performs
}
\details{
crda
}
\examples{
crda(X,y)
crda(X,y, Xt=Xt)
crda(X,y, q=1,Xt=Xt)
crda(X,y, q=2,Xt=Xt,centerX=TRUE)
}
\seealso{
\code{\link{crda.regparam}}, \code{\link{crda.cv}}
}
\author{
Muhammad Naveed Tabassum and Esa Ollila, 2018
}
