% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crda.setup3.R
\name{crda.setup3}
\alias{crda.setup3}
\title{crda.setup3}
\usage{
crda.setup3()
}
\value{
X,y, Xt,yt, truepos.
}
\description{
This function generates partially synthetic data for setup 3 of the CRDA-paper.
}
\examples{
crda.setup3()
}
